fix: ts types
