+++ src/index.ts
@@ -1,3 +1,4 @@
+export { Store, StoreProp } from './types';
 export { Provider } from './Provider';
 export { connect } from './connect';
 export { create } from './create';
