+++ src/types.tsx
@@ -19,10 +19,8 @@
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 3.0
 
-import {
-  ComponentType,
-  NamedExoticComponent
-} from 'react';
+import { ComponentType, NamedExoticComponent } from 'react';
+import { NonReactStatics } from 'hoist-non-react-statics';
 
 /**
  * This interface can be augmented by users to add default types for the root state when
@@ -33,22 +31,15 @@ import {
 // tslint:disable-next-line:no-empty-interface
 export interface DefaultRootState {}
 
-
 export interface Store<S = {}> {
   setState: (state: Partial<S>) => void;
   getState: () => S;
   subscribe: (listener: () => void) => () => void;
 }
 
-// Omit taken from https://www.typescriptlang.org/docs/handbook/release-notes/typescript-2-8.html
-export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
-
-export interface StoreProp {
-  store: Store;
+export interface StoreProp<S = {}> {
+  store: Store<S>;
 }
-
-export type AdvancedComponentDecorator<TProps, TOwnProps> =
-  (component: ComponentType<TProps>) => NamedExoticComponent<TOwnProps>;
 
 /**
 * A property P will be present if:
@@ -90,95 +81,10 @@ export type Shared<
 // Infers prop type from component C
 export type GetProps<C> = C extends ComponentType<infer P> ? P : never;
 
-// Applies LibraryManagedAttributes (proper handling of defaultProps
-// and propTypes), as well as defines WrappedComponent.
 export type ConnectedComponent<
-  C extends ComponentType<any>,
-  P
-> = NamedExoticComponent<JSX.LibraryManagedAttributes<C, P>> & {
+    C extends ComponentType<any>,
+    T,
+    P
+> = NamedExoticComponent<JSX.LibraryManagedAttributes<C, Omit<GetProps<C>, keyof Shared<T, GetProps<C>>> & P>> & NonReactStatics<C> & {
   WrappedComponent: C;
 };
-
-// Injects props and removes them from the prop requirements.
-// Will not pass through the injected props if they are passed in during
-// render. Also adds new prop requirements from TNeedsProps.
-export type InferableComponentEnhancerWithProps<TInjectedProps, TNeedsProps> =
-  <C extends ComponentType<Matching<TInjectedProps, GetProps<C>>>>(
-      component: C
-  ) => ConnectedComponent<C, Omit<GetProps<C>, keyof Shared<TInjectedProps, GetProps<C>>> & TNeedsProps>;
-
-// Injects props and removes them from the prop requirements.
-// Will not pass through the injected props if they are passed in during
-// render.
-export type InferableComponentEnhancer<TInjectedProps> =
-  InferableComponentEnhancerWithProps<TInjectedProps, {}>;
-
-export interface Connect {
-  // tslint:disable:no-unnecessary-generics
-  (): InferableComponentEnhancer<StoreProp>;
-
-  <TStateProps = {}, TOwnProps = {}, State = DefaultRootState>(
-      mapStateToProps: MapStateToPropsParam<TStateProps, TOwnProps, State>,
-      options?: Options,
-  ): InferableComponentEnhancerWithProps<TStateProps & StoreProp, TOwnProps>;
-  // tslint:enable:no-unnecessary-generics
-}
-
-/**
-* Infers the type of props that a connector will inject into a component.
-*/
-export interface ConnectProps {
-  miniStoreForwardedRef: React.Ref<any>;
-}
-
-export interface ConnectedState<TStateProps = {}, Store = {}, TOwnProps = {}> {
-  subscribed: TStateProps;
-  store: Store;
-  props: TOwnProps,
-}
-
-export type MapStateToProps<TStateProps, TOwnProps, State = DefaultRootState> =
-  (state: State, ownProps: TOwnProps) => TStateProps;
-
-export type MapStateToPropsFactory<TStateProps, TOwnProps, State = DefaultRootState> =
-  (initialState: State, ownProps: TOwnProps) => MapStateToProps<TStateProps, TOwnProps, State>;
-
-export type MapStateToPropsParam<TStateProps, TOwnProps, State = DefaultRootState> =
-  MapStateToPropsFactory<TStateProps, TOwnProps, State> | MapStateToProps<TStateProps, TOwnProps, State> | null | undefined;
-
-export type MergeProps<TStateProps, TDispatchProps, TOwnProps, TMergedProps> =
-  (stateProps: TStateProps, dispatchProps: TDispatchProps, ownProps: TOwnProps) => TMergedProps;
-
-export interface Options {
-  /**
-   * If true, use React's forwardRef to expose a ref of the wrapped component
-   *
-   * @default false
-   */
-  forwardRef?: boolean;
-}
-;
-
-export interface ProviderProps {
-  store: Store;
-}
-
-/**
-* This interface allows you to easily create a hook that is properly typed for your
-* store's root state.
-*
-* @example
-*
-* interface RootState {
-*   property: string;
-* }
-*
-* const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector;
-*
-*/
-export interface TypedUseSelectorHook<TState> {
-  <TSelected>(
-      selector: (state: TState) => TSelected,
-      equalityFn?: (left: TSelected, right: TSelected) => boolean
-  ): TSelected;
-}
